LayoutTests/media/audio-play.html

<!DOCTYPE html>
<html>
<head>
    <title>'play' event</title>
    <script src=media-file.js></script>
    <script src=video-test.js></script>

    <script>
        function start() {
            mediaElement = new Audio();
            waitForEvent('loadedmetadata');
            waitForEvent('play', function() { endTest(); });
            run("mediaElement.src = findMediaFile('audio', 'content/test')");
            run("mediaElement.play()");
        }
    </script>
</head>

<body onload="start()">
    <p>Test that a 'play' event is fired when an audio element plays.</p>
</body>
</html>
